qemu: depend on the class-target OVMF at runtime
We use qemu to run our cross-compiled (class-target) LUV images. Thus,
we need an OVMF binary that is also cross-compiled.

Plus, even though we require the OVMF binary only when running QEMU and
this would imply an RDEPENDS, both qemu and ovmf don't go into our
filesystem. Thus, we keep the DEPENDS so that a build of qemu keeps
triggers a build of ovmf.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>